Extract the rewind code from SaveState.cpp, to make it more managable.
