+++ Core/SaveState.h
@@ -33,12 +33,7 @@ namespace SaveState {
 	};
 	typedef std::function<void(Status status, std::string_view message)> Callback;
 
-	static const char * const STATE_EXTENSION = "ppst";
 	static const char * const SCREENSHOT_EXTENSION = "jpg";
-	static const char * const UNDO_STATE_EXTENSION = "undo.ppst";
-	static const char * const UNDO_SCREENSHOT_EXTENSION = "undo.jpg";
-
-	static const char * const LOAD_UNDO_NAME = "load_undo.ppst";
 
 	void Init();
 	void Shutdown();
