+++ UWP/CoreUWP/CoreUWP.vcxproj
@@ -309,6 +309,7 @@
     <ClInclude Include="..\..\Core\HLE\Plugins.h" />
     <ClInclude Include="..\..\Core\RetroAchievements.h" />
     <ClInclude Include="..\..\Core\SaveState.h" />
+    <ClInclude Include="..\..\Core\SaveStateRewind.h" />
     <ClInclude Include="..\..\Core\Screenshot.h" />
     <ClInclude Include="..\..\Core\System.h" />
     <ClInclude Include="..\..\Core\TiltEventProcessor.h" />
@@ -616,6 +617,7 @@
     <ClCompile Include="..\..\Core\HLE\Plugins.cpp" />
     <ClCompile Include="..\..\Core\RetroAchievements.cpp" />
     <ClCompile Include="..\..\Core\SaveState.cpp" />
+    <ClCompile Include="..\..\Core\SaveStateRewind.cpp" />
     <ClCompile Include="..\..\Core\Screenshot.cpp" />
     <ClCompile Include="..\..\Core\System.cpp" />
     <ClCompile Include="..\..\Core\TiltEventProcessor.cpp" />
