+++ libretro/Makefile.common
@@ -879,6 +879,7 @@ SOURCES_CXX += \
 	       $(COREDIR)/Reporting.cpp \
 	       $(COREDIR)/RetroAchievements.cpp \
 	       $(COREDIR)/SaveState.cpp \
+	       $(COREDIR)/SaveStateRewind.cpp \
 	       $(COREDIR)/Screenshot.cpp \
 	       $(COREDIR)/System.cpp \
 	       $(COREDIR)/Util/AtracTrack.cpp \
